+++ client/src/components/table/ColumnFilter.test.tsx
@@ -10,7 +10,7 @@ import { createColumnHelper } from "@tanstack/react-table";
 import { ColumnFilter } from "./ColumnFilter";
 import { isAfter, isBefore, isSameDay } from "date-fns";
 import { formatDateForDisplay } from "util/formatDate";
-import { clearWebStorage } from "hooks/useLocalStorage";
+import { clearWebStorage } from "hooks";
 
 const columnHelper = createColumnHelper<TestType>();
