+++ client/src/components/table/KeywordSearch.test.tsx
@@ -7,7 +7,7 @@ import userEvent from "@testing-library/user-event";
 import { Table } from "./Table";
 import { TestType, testTableData } from "./Table.test";
 import { highlightCell, KeywordSearch, TEST_IDS } from "./KeywordSearch";
-import { clearWebStorage } from "hooks/useLocalStorage";
+import { clearWebStorage } from "hooks";
 import { createColumnHelper } from "@tanstack/react-table";
 
 const columnHelper = createColumnHelper<TestType>();
