+++ client/src/components/table/tables/DocumentTable.test.tsx
@@ -2,7 +2,7 @@ import React from "react";
 
 import { ALL_MOCKS } from "mock-data/index";
 import { beforeEach, describe, expect, it } from "vitest";
-import { clearWebStorage } from "hooks/useLocalStorage";
+import { clearWebStorage } from "hooks";
 
 import { MockedProvider } from "@apollo/client/testing";
 import { fireEvent, render, screen, waitFor, within } from "@testing-library/react";
